requirements.txt
boto==2.42.0
boto3==1.4.1
botocore==1.4.62
docutils==0.12
futures==3.0.5
jmespath==0.9.0
python-dateutil==2.5.3
python-lambda-local==0.1.2
requests==2.11.1
s3transfer==0.1.8
six==1.10.0
wheel==0.24.0
